fix(epoch): contain a bad tree, align the event type, guard the boundary math
Four smaller points from the same review.

A batch that somehow reached a block-height scheduler aborted the entire sweep
scan by returning an error from the Apply callback, so one misconfigured tree
would stop every other batch from being swept. Settings validation already
refuses that pairing, so it is an invariant violation: log it at error level and
skip that tree instead.

BatchStartedEvent.BatchExpiryDate was uint32 internally against an int64 on the
wire. Nothing truncates today - AbsoluteLocktime is itself a uint32 - but the
asymmetry is a trap for anything that later reads the field back.

BoundaryAfter multiplied the epoch count by the length with no overflow check.
Unreachable with a validated anchor, but wrapping would produce a boundary in
the past, which is the one failure here that would not look wrong. It now
returns the zero time, which ExpiryFor already rejects through its range check.

Added a startup line naming the current boundary and the admission window, since
the default anchor is a fixed date in the past by design and an operator has no
other way to see which dates their batches land on. Also flagged in the e2e file
that a failure there should be read as design-level first.

Reported by arkana-ai-bot on #1171.

+++ internal/core/application/service_event.go
@@ -21,7 +21,10 @@ type BatchStarted struct {
 	// legacy batch. Carried separately from BatchExpiry because an old client
 	// reads that field as a relative locktime and would fail BIP68 encoding on an
 	// absolute timestamp - safe, but opaque.
-	BatchExpiryDate uint32
+	//
+	// int64 to match the wire type, so the value cannot be narrowed on its way to
+	// the proto. AbsoluteLocktime is a uint32, so it always fits.
+	BatchExpiryDate int64
 	// UnrollGrace mirrors BatchExpiry for epoch batches, named for what it is.
 	UnrollGrace uint32
 }
